requirements.txt
customtkinter
pillow
google-generativeai
pystray
winotify
pyinstaller
